+++ sqlathanor/declarative.py
@@ -2331,7 +2331,8 @@ class BaseModel(object):
                     double_wrapper_character_when_nested = False,
                     escape_character = "\\",
                     line_terminator = '\r\n'):
-        r"""Retrieve a CSV string with all object attributes serialized.
+        r"""Retrieve a :term:`CSV <Comma-Separated Value (CSV)>` representation of
+        the object, *with all attributes* serialized regardless of configuration.
 
         .. caution::
 
@@ -2414,7 +2415,9 @@ class BaseModel(object):
                      current_nesting = 0,
                      serialize_function = None,
                      **kwargs):
-        """Return a JSON representation of the object.
+        """Return a :term:`JSON <JavaScript Object Notation (JSON)>`
+        representation of the object, *with all attributes* regardless of
+        configuration.
 
         .. caution::
 
@@ -2481,6 +2484,81 @@ class BaseModel(object):
         as_json = serialize_function(as_dict, **kwargs)
 
         return as_json
+
+    def dump_to_yaml(self,
+                     max_nesting = 0,
+                     current_nesting = 0,
+                     serialize_function = None,
+                     **kwargs):
+        """Return a :term:`YAML <YAML Ain't a Markup Language (YAML)>`
+        representation of the object *with all attributes*, regardless of
+        configuration.
+
+        .. caution::
+
+          Nested objects (such as :term:`relationships <relationship>` or
+          :term:`association proxies <association proxy>`) will **not**
+          be serialized.
+
+        :param max_nesting: The maximum number of levels that the resulting
+          object can be nested. If set to ``0``, will not nest other serializable
+          objects. Defaults to ``0``.
+        :type max_nesting: :class:`int <python:int>`
+
+        :param current_nesting: The current nesting level at which the
+          representation will reside. Defaults to ``0``.
+        :type current_nesting: :class:`int <python:int>`
+
+        :param serialize_function: Optionally override the default YAML serializer.
+          Defaults to :obj:`None <python:None>`, which calls the default ``yaml.dump()``
+          function from the `PyYAML <https://github.com/yaml/pyyaml>`_ library.
+
+          .. note::
+
+            Use the ``serialize_function`` parameter to override the default
+            YAML serializer.
+
+            A valid ``serialize_function`` is expected to
+            accept a single :class:`dict <python:dict>` and return a
+            :class:`str <python:str>`, similar to ``yaml.dump()``.
+
+            If you wish to pass additional arguments to your ``serialize_function``
+            pass them as keyword arguments (in ``kwargs``).
+
+        :type serialize_function: callable / :obj:`None <python:None>`
+
+        :param kwargs: Optional keyword parameters that are passed to the
+          YAML serializer function. By default, these are options which are passed
+          to ``yaml.dump()``.
+        :type kwargs: keyword arguments
+
+        :returns: A :class:`str <python:str>` with the JSON representation of the
+          object.
+        :rtype: :class:`str <python:str>`
+
+        :raises SerializableAttributeError: if attributes is empty
+        :raises MaximumNestingExceededError: if ``current_nesting`` is greater
+          than ``max_nesting``
+        :raises MaximumNestingExceededWarning: if an attribute requires nesting
+          beyond ``max_nesting``
+
+        """
+        if serialize_function is None:
+            serialize_function = yaml.dump
+        else:
+            if checkers.is_callable(serialize_function) is False:
+                raise ValueError(
+                    'serialize_function (%s) is not callable' % serialize_function
+                )
+
+        as_dict = self._to_dict('yaml',
+                                max_nesting = max_nesting,
+                                current_nesting = current_nesting,
+                                is_dumping = True)
+
+        as_yaml = serialize_function(as_dict, **kwargs)
+
+        return as_yaml
 
     @classmethod
     def _parse_dict(cls,