+++ CHANGES.rst
@@ -8,6 +8,7 @@ New Features
 
 * #35: Added ``BaseModel.dump_to_csv()``
 * #35: Added ``BaseModel.dump_to_json()``
+* #35: Added ``BaseModel.dump_to_yaml()``
 
 -----------